backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
tensorflow-cpu==2.20.0
numpy
opencv-python-headless
Pillow
firebase-admin
h5py
gdown